Доработка логирования crond, чтобы информация о запускаемых задачах показывалать только при уровне логирования DEBUG

+++ files/etc/my_init.d/000_startup.sh
@@ -1,3 +1,3 @@
 #!/usr/bin/env bash
 
-rm /tmp/SYSLOG_STARTED || true
+[ -f "/tmp/SYSLOG_STARTED" ] && rm /tmp/SYSLOG_STARTED || true
